content/media/webaudio/test/test_audioParamSetCurveAtTime.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test AudioParam.linearRampToValue</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  SpecialPowers.setBoolPref("media.webaudio.enabled", true);

  var context = new AudioContext();

  var T0 = 0;
  var duration = 1024 / context.sampleRate;

  var sourceBuffer = context.createBuffer(1, 2048, context.sampleRate);
  for (var i = 0; i < 2048; ++i) {
    sourceBuffer.getChannelData(0)[i] = 1;
  }
  var curve = new Float32Array(100);
  for (var i = 0; i < 100; ++i) {
    curve[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
  }
  var expectedBuffer = context.createBuffer(1, 2048, context.sampleRate);
  for (var i = 0; i < 2048; ++i) {
    var t = i / context.sampleRate;
    expectedBuffer.getChannelData(0)[i] = curve[Math.min(99, Math.floor(100 * Math.min(1.0, (t - T0) / duration)))];
  }

  var destination = context.destination;

  var source = context.createBufferSource();
  source.buffer = sourceBuffer;

  var gain = context.createGain();
  gain.gain.setValueCurveAtTime(curve, T0, duration);

  var sp = context.createScriptProcessor(2048, 1);
  source.connect(gain);
  gain.connect(sp);
  sp.connect(destination);

  source.start(0);
  sp.onaudioprocess = function(e) {
    is(e.inputBuffer.numberOfChannels, 1, "Correct input channel count");
    compareBuffers(e.inputBuffer.getChannelData(0), expectedBuffer.getChannelData(0));

    sp.onaudioprocess = null;

    SpecialPowers.clearUserPref("media.webaudio.enabled");
    SimpleTest.finish();
  };
});

</script>
</pre>
</body>
</html>
